Use homepage.tsx file for index, redirect accordingly
